soap/requirements.txt
azure-eventgrid==1.3.0
azure-storage-file-datalake==12.1.2
jupyter==1.0.0
lxml==4.6.1
numpy==1.19.4
pandas==1.1.4
requests==2.25.0
zeep==4.0.0